+++ pywxdump/db/dbMedia.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-#
 # -------------------------------------------------------------------------------
-# Name:         MediaMSG_parsing.py
-# Description:  
+# Name:         MediaMSG.py
+# Description:  负责处理语音数据库
 # Author:       xaoyaoo
 # Date:         2024/04/15
 # -------------------------------------------------------------------------------
@@ -9,14 +9,23 @@ from .dbbase import DatabaseBase
 from .utils import silk2audio
 
 
-class ParsingMediaMSG(DatabaseBase):
+class MediaHandler(DatabaseBase):
     _class_name = "MediaMSG"
-    def __init__(self, db_path):
-        super().__init__(db_path)
+    Media_required_tables = ["Media"]
+
+    def Media_add_index(self):
+        """
+        添加索引, 加快查询速度
+        """
+        if self.tables_exist("Media"):
+            self.execute("CREATE INDEX IF NOT EXISTS MsgSvrID ON Media(Reserved0)")
 
     def get_audio(self, MsgSvrID, is_play=False, is_wave=False, save_path=None, rate=24000):
+        if not self.tables_exist("Media"):
+            return False
+
         sql = "select Buf from Media where Reserved0=? "
-        DBdata = self.execute_sql(sql, (MsgSvrID,))
+        DBdata = self.execute(sql, (MsgSvrID,))
         if not DBdata:
             return False
         if len(DBdata) == 0: